wip: implementation of mvp (see below) design pat
http://de.wikipedia.org/wiki/Model_View_Controller
this design pattern seems reasonable for the organization of the client
gui. it clearly seperates the data preservation from display on screen.
This is handled in three classes: a view class, which contains all
objects regarding the display on screen, a model class for the
preservation of data and a controller class, that fits the two together.